requirements.txt
opencv-python
numpy
mediapipe
pygame
screeninfo
PyQt5
pyserial